# docs/notifications/openapi.yaml
openapi: 3.0.3
info:
  title: Massava Notification API
  description: |
    Comprehensive API documentation for the Massava notification system.

    The notification system provides:
    - Real-time in-app notifications
    - Push notifications (iOS, Android, Web)
    - Email notifications (instant and digest)
    - Server-Sent Events (SSE) for real-time updates
    - Customizable user preferences
    - Device token management for push notifications

    All endpoints require authentication via Bearer token (JWT).
  version: 1.0.0
  contact:
    name: Massava Development Team
    url: https://github.com/roman/massava
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.massava.com
    description: Production API
  - url: https://staging-api.massava.com
    description: Staging API
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication endpoint
  schemas:
...
    DigestFrequency:
      type: string
      enum:
        - DAILY
        - WEEKLY
      description: |
        Frequency for email digest notifications.
        - DAILY: Digest sent daily
        - WEEKLY: Digest sent weekly
    TypePreference:
      type: object
      required:
        - push
        - email
        - inApp
      properties:
        push:
          type: boolean
          description: Whether to send push notifications for this type
        email:
          type: string
          enum:
            - instant
            - digest
            - off
          description: |
            Email delivery mode.
            - instant: Send email immediately
            - digest: Batch in daily/weekly digest
            - off: Don't send email
        inApp:
          type: boolean
          description: Whether to show in-app notification
      example:
        push: true
        email: instant
        inApp: true
    NotificationPreferences:
      type: object
      required:
        - id
        - userId
        - pushEnabled
        - emailEnabled
        - inAppEnabled
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: Preferences identifier
        userId:
          type: string
          description: User ID
        pushEnabled:
          type: boolean
          description: Global enable/disable for push notifications
        emailEnabled:
          type: boolean
          description: Global enable/disable for email notifications
        inAppEnabled:
          type: boolean
          description: Global enable/disable for in-app notifications
        quietHoursEnabled:
          type: boolean
          description: Whether quiet hours are enabled
        quietHoursStart:
          type: string
          nullable: true
          pattern: '^\d{2}:\d{2}$'
          description: Quiet hours start time (HH:MM format)
          example: '22:00'
        quietHoursEnd:
          type: string
          nullable: true
          pattern: '^\d{2}:\d{2}$'
          description: Quiet hours end time (HH:MM format)
          example: '08:00'
        timezone:
          type: string
          description: User's timezone for quiet hours
          example: Europe/Berlin
        typePreferences:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/TypePreference'
          nullable: true
          description: Per-type notification preferences
        emailDigestEnabled:
          type: boolean
          description: Whether email digest is enabled
        digestFrequency:
          $ref: '#/components/schemas/DigestFrequency'
        digestTime:
          type: string
          nullable: true
          pattern: '^\d{2}:\d{2}$'
          description: Time to send email digest (HH:MM format)
          example: '09:00'
        language:
          type: string
          description: Language for notification content
          example: de
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      example:
        id: pref123
        userId: user123
        pushEnabled: true
        emailEnabled: true
        inAppEnabled: true
        quietHoursEnabled: true
        quietHoursStart: '22:00'
        quietHoursEnd: '08:00'
        timezone: Europe/Berlin
        emailDigestEnabled: true
        digestFrequency: DAILY
        digestTime: '09:00'
        language: de
        typePreferences:
          BOOKING_CONFIRMED:
            push: true
            email: instant
            inApp: true
          STUDIO_PROMOTION:
            push: false
            email: digest
            inApp: false
        createdAt: '2025-01-01T00:00:00Z'
        updatedAt: '2025-01-15T10:00:00Z'
    UpdateNotificationPreferencesRequest:
      type: object
      properties:
        pushEnabled:
          type: boolean
        emailEnabled:
          type: boolean
        inAppEnabled:
          type: boolean
        quietHoursEnabled:
          type: boolean
        quietHoursStart:
          type: string
          pattern: '^\d{2}:\d{2}$'
          nullable: true
        quietHoursEnd:
          type: string
          pattern: '^\d{2}:\d{2}$'
          nullable: true
        timezone:
          type: string
        typePreferences:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/TypePreference'
        emailDigestEnabled:
          type: boolean
        digestFrequency:
          $ref: '#/components/schemas/DigestFrequency'
        digestTime:
          type: string
          pattern: '^\d{2}:\d{2}$'
          nullable: true
        language:
          type: string
      example:
        pushEnabled: true
        emailEnabled: false
        quietHoursEnabled: true
        quietHoursStart: '22:00'
        quietHoursEnd: '08:00'
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        message:
          type: string
          nullable: true
          description: Additional error details
        details:
          type: object
          nullable: true
          description: Structured error details (e.g., validation errors)
        statusCode:
          type: integer
          description: HTTP status code
      example:
        error: Invalid request
        message: Request validation failed
        details:
          token:
            - Invalid device token format
    UnauthorizedError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          enum:
            - Unauthorized
            - Nicht autorisiert
        message:
          type: string
          nullable: true
...
  responses:
    UnauthorizedResponse:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UnauthorizedError'
...
    BadRequestResponse:
      description: Invalid request parameters or body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerErrorResponse:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Internal server error
            statusCode: 500
security:
  - BearerAuth: []
paths:
...
  /api/notifications/preferences:
    get:
      tags:
        - Preferences
      summary: Get notification preferences
      operationId: getPreferences
      description: |
        Retrieve notification preferences for the authenticated user.

        Returns default preferences if none have been set yet.
      responses:
        '200':
          description: Notification preferences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPreferences'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'

    patch:
      tags:
        - Preferences
      summary: Update notification preferences
      operationId: updatePreferences
      description: |
        Update notification preferences for the authenticated user.

        Supports partial updates - only changed fields need to be provided.

        **Features**:
        - Global channel enable/disable (push, email, in-app)
        - Quiet hours configuration (time range + timezone)
        - Per-notification-type preferences (channel selection + email mode)
        - Email digest settings (frequency + delivery time)
        - Language selection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNotificationPreferencesRequest'
      responses:
        '200':
          description: Preferences updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPreferences'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
...